fix: resolve golangci-lint path via GOBIN/GOPATH in Makefile
Add GOLANGCI_LINT variable following the same GOBIN/GOPATH fallback
pattern used for AIR_BIN. The pre-commit hook runs in a non-interactive
shell without ~/go/bin on PATH, so command -v golangci-lint failed.
Use the resolved path in lint-golangci, lint-golangci-ci, and
nilaway-golangci-build targets.

Also apply mdformat auto-formatting (80-char wrap) to tracked
Markdown files.
